Start work on test for archiveWorkspace - Also add stubs for some needed alias-related methods

+++ app/src/main/kotlin/mensagod/dbcmds/Workspace.kt
@@ -1,5 +1,7 @@
 package mensagod.dbcmds
 
+import keznacl.BadValueException
+import keznacl.onTrue
 import keznacl.toFailure
 import keznacl.toSuccess
 import libkeycard.Domain
@@ -9,6 +11,7 @@ import libmensago.NotConnectedException
 import libmensago.ResourceExistsException
 import mensagod.DBConn
 import mensagod.DatabaseCorruptionException
+import mensagod.UnauthorizedException
 
 enum class WorkspaceType {
     Individual,
@@ -104,11 +107,33 @@ fun addWorkspace(
 /**
  * Archives a workspace. Mensago workspaces are never deleted, as they are not intended to be
  * reused. Archiving is a permanent change, so don't do this unless you're sure!
+ *
+ * @exception NotConnectedException if not connected to the database
+ * @exception java.sql.SQLException for database problems, most likely either with your query or
+ * with the connection
+ * @exception BadValueException Returned if given an alias
+ * @exception UnauthorizedException Returned if you try to archive the admin account
+ * @exception DatabaseCorruptionException Returned if there's a problem with the data in the db
  */
 fun archiveWorkspace(db: DBConn, wid: RandomID): Throwable? {
+    isAlias(db, wid).getOrElse { return it }.onTrue { return BadValueException() }
+    val rs = db.query("SELECT wid FROM workspaces WHERE uid='admin'").getOrElse { return it }
+    if (!rs.next()) return DatabaseCorruptionException("admin workspace missing")
+    val adminWID = RandomID.fromString(rs.getString("wid"))
+        ?: return DatabaseCorruptionException("invalid WID for admin workspace")
+    if (adminWID == wid) return UnauthorizedException("admin can't be archived")
+
+    val aliases = getAliases(db, wid).getOrElse { return it }
+    if (aliases.isNotEmpty()) {
+        db.execute("DELETE FROM aliases WHERE wid=?", wid).getOrElse { return it }
+        aliases.forEach {
+            db.execute("DELETE FROM aliases WHERE alias=?", it).getOrElse { return it }
+        }
+    }
 
     db.execute("UPDATE workspaces SET password='-',status='archived' WHERE wid=?", wid)
         .getOrElse { return it }
+    db.execute("DELETE FROM keycards WHERE wid=?", wid).getOrElse { return it }
     db.execute("DELETE FROM iwkspc_folders WHERE wid=?", wid).getOrElse { return it }
     db.execute("DELETE FROM quotas WHERE wid=?", wid).getOrElse { return it }
     db.execute("DELETE FROM updates WHERE wid=?", wid).getOrElse { return it }
@@ -142,6 +167,13 @@ fun checkWorkspace(db: DBConn, wid: RandomID): Result<WorkspaceStatus?> {
 }
 
 /**
+ * Returns all the aliases for a workspace ID
+ */
+fun getAliases(db: DBConn, wid: RandomID): Result<List<RandomID>> {
+    TODO("Implement getAliases($db, $wid)")
+}
+
+/**
  * Returns a workspace ID which does not already exist in the system.
  */
 fun getFreeWID(db: DBConn): Result<RandomID> {
@@ -162,13 +194,25 @@ fun getFreeWID(db: DBConn): Result<RandomID> {
  * resolveUserID attempts to return the RandomID corresponding to the specified UserID. If it does
  * not exist, null is returned.
  *
- * @throws NotConnectedException if not connected to the database
- * @throws java.sql.SQLException for database problems, most likely either with your query or with the connection
+ * @exception NotConnectedException if not connected to the database
+ * @exception java.sql.SQLException for database problems, most likely either with your query or
+ * with the connection
  */
 fun isAlias(db: DBConn, wid: RandomID): Result<Boolean> {
     val rs = db.query("SELECT alias FROM aliases WHERE wid=?", wid)
         .getOrElse { return it.toFailure() }
     return rs.next().toSuccess()
+}
+
+/**
+ * Creates an alias to an existing workspace.
+ *
+ * @exception NotConnectedException if not connected to the database
+ * @exception java.sql.SQLException for database problems, most likely either with your query or
+ * with the connection
+ */
+fun makeAlias(db: DBConn, wid: RandomID, uid: UserID?): Result<RandomID> {
+    TODO("Implement makeAlias($db, $wid, $uid")
 }
 
 /**
